requirements.txt
PyPDF2==3.0.1
langchain-community>=0.0.38
langchain-core>=0.1.53
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
openai>=1.12.0
python-dotenv==1.0.0
rich==13.7.0
faiss-cpu==1.7.4
tiktoken==0.5.2
streamlit==1.31.0 